+++ src/sequencing.rs
@@ -62,7 +62,7 @@ pub(crate) struct SequenceState {
 pub struct Sequence {
     id: usize,
     elements: Vec<SequenceElement>,
-    bars: usize,
+    bars: f32,
 }
 
 pub fn chord(notes: impl Into<Vec<i32>>) -> SequenceElement {
@@ -83,7 +83,7 @@ pub fn seq(id: usize, elements: impl Into<Vec<SequenceElement>>) -> Sequence {
     Sequence {
         id,
         elements: elements.into(),
-        bars: 1,
+        bars: 1.,
     }
 }
 
@@ -92,7 +92,7 @@ pub fn sub(elements: impl Into<Vec<SequenceElement>>) -> SequenceElement {
 }
 
 impl Sequence {
-    pub fn bars(mut self, bars: usize) -> Self {
+    pub fn bars(mut self, bars: f32) -> Self {
         self.bars = bars;
         self
     }
@@ -116,11 +116,11 @@ impl Sequence {
         self.ensure_state(state);
 
         if clock_position < state.last_clock_position {
-            state.current_bar = (state.current_bar + 1) % self.bars;
+            state.current_bar = ((state.current_bar + 1) as f32 % self.bars) as usize;
         }
         state.last_clock_position = clock_position;
 
-        let sequence_position = (state.current_bar as f32 + clock_position) / self.bars as f32;
+        let sequence_position = (state.current_bar as f32 + clock_position) / self.bars;
 
         let (chord_index, active_chord) =
             Self::find_active_chord(&self.elements, sequence_position);
